backend/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
# OR for MongoDB
# beanie==1.25.0
# motor==3.3.2

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-dotenv==1.0.0

# Payment Gateways
stripe==7.10.0
paypalcheckoutcore==1.3.0

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Email
fastapi-mail==1.4.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3
uuid==1.30

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Caching
redis==5.0.1
fastapi-cache2==0.2.1

# Logging
loguru==0.7.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Development
black==23.12.1
flake8==7.0.0
isort==5.13.2
pre-commit==3.6.0